Remove unwanted padding in math tables
When writing a display equation in the "aligned" environment, pandoc
creates a mathml table, but this gets styled by Firefox (and MathJax)
with padding between all table cells, while I only want padding between
odd-even boundaries (like in align). This uses CSS to put the padding
only where I want it. Doesn't work on MathJax, which is just another
reason why it's not a good replacement for MathML.